refactor(cli): split Oxigraph supervisor ownership
